# src/main/java/dk/sdu/mmmi/backendforfrontend/service/model/LoginRequest.java
package dk.sdu.mmmi.backendforfrontend.service.model;

import lombok.Getter;
import lombok.Setter;
import lombok.ToString;

@Getter
@Setter
@ToString
public class LoginRequest {

    private String email;

    @ToString.Exclude
    private String password;

}